Versão 2.0
Exclusão de gerenciadorDeEstoque.mwb.bak


Update Model

Criação da classe GeralDAO. A partir de agora ProdutoDAO e VendaDAO, estendem da nova classe.
Update Geral

Javadoc + Mudança do tipo de variável (código) + Redução de código
Update Geral

Javadoc + Mudança do tipo de variável (código) + Redução de código
Delete TelaLogin


Updates Janelas de Relatório

Javadoc + Redução de Código
Update CadastroNovoProduto

Mensagem de campos que não podem ficar vazios
Update Arquivo Banco de Dados

Inserção da tabela Cliente no banco de dados.
Update Produto

Update Banco de Dados + Tela de Inserção + Tela VerEstoque + Gerador de PDF
Adicionadas Classes de Clientes


Update Telas e Classes Clientes


Update Telas

Dispose ao fechar

+++ src/view/InserirEstoqueProduto.java
@@ -165,7 +165,7 @@ public class InserirEstoqueProduto extends javax.swing.JFrame {
         // TODO add your handling code here:
         if (!txt_codigoProduto.getText().isEmpty() && !txt_addEstoque.getText().isEmpty()) {
             ProdutoDAO daoProduto = new ProdutoDAO();
-            int codigo = Integer.parseInt(txt_codigoProduto.getText());
+            long codigo = Long.parseLong(txt_codigoProduto.getText());
             int addEstoque = Integer.parseInt(txt_addEstoque.getText());
             if (daoProduto.inserirEstoqueProduto(addEstoque, codigo)) {
                 JOptionPane.showMessageDialog(null, "Estoque Inserido");
@@ -186,7 +186,7 @@ public class InserirEstoqueProduto extends javax.swing.JFrame {
 
         if (!txt_codigoProduto.getText().isEmpty()) {
             ProdutoDAO daoProduto = new ProdutoDAO();
-            int codigo = Integer.parseInt(txt_codigoProduto.getText());
+            long codigo = Long.parseLong(txt_codigoProduto.getText());
             Produto produto = daoProduto.getInfoProduto(codigo);
 
             lbl_nome.setText("Nome: " + produto.getNomeProduto());
